fix(certificate-authority): decode ledger serial as uint64_t, not int64_t
Root cause of the intermittent ca_cluster_node_test "Commit timeout"
failures on CI (both property_17_issuance_survives_leader_failover and
restarted_follower_recovers_and_can_become_leader) that raising
k_command_timeout to 60s didn't fix: this was never a timing problem.

ca_ledger_entry::serial is std::uint64_t — certificate serial numbers are
non-sequential serials), so roughly half of all real serials have bit 63
set and don't fit in an int64_t. json_to_ledger_entry() and the
record_revocation command handler (include/raft/ca_state_machine.hpp)
decoded the "serial" JSON field via .to_number<std::int64_t>() instead of
.to_number<std::uint64_t>(); boost::json's to_number<T>() throws a
"not exact" error when the stored value can't be exactly represented in
the requested type. Encoding was always correct (obj["serial"] = e.serial
stores a uint64_t value with the right boost::json kind) — only decoding
had the bug.

apply_committed_entries() halts entry application on any state-machine
apply() failure (application_failure_policy::halt, the default) and never
retries that index, so once a large-serial issuance landed at some log
index, every later index silently stopped being applied too —
commit_index kept advancing via ordinary replication (visible in the
leader's own heartbeat logs), but nothing past the failed entry was ever
applied or notified, so every later client's submit_command() future for
those app cations sat until commit_waiter's periodic sweep timed it out.
This explains the intermittent, seemingly timing-dependent failures
exactly: whether a run failed depended entirely on whether that run's
randomly-generated certificate serial happened to have its high bit set,
not on host load or timeout size.

Diagnosed by tracing CI's captured DEBUG logs: "Advanced commit index"
correctly reached the affected index while "Applying log entry to state
machine" for an EARLIER index in the same batch logged "ERROR: Failed to
apply log entry to state machine ... error=not exact ... to_number ...",
confirming apply had halted partway through a multi-entry catch-up batch.

Reproduced directly and deterministically via a new regression test
(tests/ca_state_machine_unit_test.cpp,
record_issuance_survives_serial_exceeding_int64_max) using
std::numeric_limits<uint64_t>::max() as the serial — this is why the bug
escaped every prior unit test: they all use small, hand-picked serials
(1, 2, 7, 42, 999) that never exercise the high bit. Verified the new test
reproduces boost::json's exact "not exact" error against the pre-fix code,
then passes against the fix; also verified with 13 consecutive
ca_cluster_node_test runs post-fix (this test's own certificates use
randomly-generated serials, so repeated runs exercise many different
values).

Reverts the k_command_timeout/test-timeout increases from the previous
commit only in spirit, not in code — those remain as reasonable, real
headroom against genuine host contention (a separate, legitimate concern
this investigation also surfaced), just not the fix for this bug.

+++ include/raft/ca_state_machine.hpp
@@ -317,7 +317,7 @@ namespace ca_state_machine_detail {
 
 [[nodiscard]] inline auto json_to_ledger_entry(const boost::json::object& obj) -> ca_ledger_entry {
     ca_ledger_entry e;
-    e.serial = static_cast<std::uint64_t>(obj.at("serial").to_number<std::int64_t>());
+    e.serial = obj.at("serial").to_number<std::uint64_t>();
     e.subject = std::string(obj.at("subject").as_string());
     e.dns_names = to_string_vector(obj.at("dns_names").as_array());
     e.ip_addresses = to_string_vector(obj.at("ip_addresses").as_array());
@@ -436,8 +436,7 @@ public:
                 return {};
             }
             case ca_command_type::record_revocation: {
-                auto serial =
-                    static_cast<std::uint64_t>(obj.at("serial").to_number<std::int64_t>());
+                auto serial = obj.at("serial").to_number<std::uint64_t>();
                 auto it =
                     std::find_if(_ledger.begin(), _ledger.end(),
                                  [serial](const ca_ledger_entry& e) { return e.serial == serial; });
